+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The resulting `.d.ts` are **only** compatible with typescript versions `>= 3.6`.
 
 ### Fixed
 
+- Fixed adding or removing edges or nodes with the api not always triggering an event
 - Fixed the result of `calculateBoundingRect` sometimes having negative width or height
 - Fixed `GroupBehaviour` interface not allowing additional properties